lanenet-enet-hnet/requirements.txt
easydict==1.6
glog==0.3.1
gast==0.2.2
loguru
tqdm
matplotlib
cmake
opencv-python==4.5.3.56 
numpy
scikit_learn
PyYAML
tensorflow_gpu==1.15.0
ipykernel
 